backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
pydantic-settings==2.13.1
python-multipart==0.0.20
supabase==2.28.2
anthropic==0.85.0
PyJWT==2.12.1
PyMuPDF==1.27.2
requests==2.32.3
psycopg[binary,pool]
pgvector==0.4.2

# Optional: only needed if you switch back to Pydantic EmailStr fields.
email-validator==2.3.0

